# src/main/webapp/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Policy Machine API"
  },
  "host": "localhost:8080",
  "basePath": "/pmapi/",
...
  "schemes": [
    "http"
  ],
  "definitions": {
...
    "Prohibition": {
      "properties": {
        "subject": {
          "$ref": "#/definitions/ProhibitionSubject"
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProhibitionResource"
          }
        }
      }
    },
    "ProhibitionResource": {
      "properties": {
        "compliment": {
          "type": "boolean",
          "example": true
        },
        "resourceId": {
          "type": "integer",
          "example": 12345
        }
      }
    },
...
    "ProhibitionSubject": {
      "properties": {
        "subjectId": {
          "type": "integer",
          "example": 100
        },
        "subjectType": {
          "type": "string",
          "enum": [
            "U",
            "UA",
            "P"
          ],
          "example": "U"
        }
      }
    },
...
    "AddOperationsToProhibitionRequest": {
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "r",
            "w"
          ]
        }
      }
    },
    "AddResourceToProhibitionRequest": {
      "properties": {
        "compliment": {
          "type": "boolean",
          "example": true
        },
        "resourceId": {
          "type": "integer",
          "example": 12345
        }
      }
    },
...
    "AssociationRequest": {
      "properties": {
        "inherit": {
          "type": "boolean",
          "example": true
        },
        "ops": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "r",
            "w"
          ]
        },
        "targetId": {
          "type": "integer",
          "example": 12345
        },
        "uaId": {
          "type": "integer",
          "example": 102
        }
      }
    },
...
    "CreateProhibitionRequest": {
      "properties": {
        "intersection": {
          "type": "boolean",
          "example": true
        },
        "name": {
          "type": "string",
          "example": "denyName"
        },
        "operations": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "r"
          ]
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProhibitionResource"
          }
        },
        "subject": {
          "$ref": "#/definitions/ProhibitionSubject"
        }
      }
    },
    "PostSubjectToProhibitionRequest": {
      "properties": {
        "subjectId": {
          "type": "integer",
          "example": 12345
        },
        "subjectType": {
          "type": "string",
          "example": "U"
        }
      }
    },
...
  },
  "paths": {
...
    "/access/resources/{targetId}": {
      "get": {
        "tags": [
          "access"
        ],
        "description": "Get the permitted operations on a resource for a specified user, or if none, all users",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "targetId",
            "description": "The ID of the resource",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "name": "userId",
            "description": "the ID of the user",
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the operations each user has on the target node",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "NodeNotFoundException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "access"
        ],
        "description": "Update an association",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "targetId",
            "description": "The ID of the target resource",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "name": "add",
            "description": "If true, add the operations in the request to the association, otherwise remove the operations from the assocaiation.  If no Association exists, an error is returned.",
            "in": "query",
            "required": false,
            "default": true
          },
          {
            "name": "body",
            "description": "The request object containing the User Attribute ID and Target ID, as well as the operations and/or a boolean inherit value. The operations and inherit value are not required, but the User Attribute ID and the TargetId are required.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AssociationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated association"
          },
          "400": {
            "description": "NodeNotFoundException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "access"
        ],
        "description": "Delete an association between a User Attribute and a target Node",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "targetId",
            "description": "The target ID",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "name": "uaId",
            "description": "The ID of the User Attribute",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Access successfully removed"
          },
          "400": {
            "description": "NodeNotFoundException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/access/resources/{targetId}/prohibitions": {
      "get": {
        "tags": [
          "access"
        ],
        "description": "Get the operations that are prohibited on a node for a user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "targetId",
            "description": "The ID of the node to get the prohibited operations",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "name": "userId",
            "description": "The ID of the user",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of operations that are prohibited for a user on a node",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "NodeNotFoundException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions": {
      "post": {
        "tags": [
          "prohibitions"
        ],
        "description": "Create a new Prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "description": "A request object containing the name, boolean intersection value, operations, resources, and subject for a new Prohibition.  The name is the only parameter that is requred.  The others will be added to the prohibition if they are present.  The default inherit value is true.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProhibitionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Prohibition was successfully created"
          },
          "400": {
            "description": "ProhibitionNameExistException, ProhibitionDoesNotExistException(When adding resources and a subject to the prohibition), NodeNotFoundException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}": {
      "get": {
        "tags": [
          "prohibitions"
        ],
        "description": "Get a Prohibition with the given name",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition to search for",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A prohibition object with the given name",
            "schema": {
              "$ref": "#/definitions/Prohibition"
            }
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "prohibitions"
        ],
        "description": "Delete a prohibition with the given name",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Prohibition was deleted successfully"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}/operations": {
      "get": {
        "tags": [
          "prohibitions"
        ],
        "description": "Get the operations for a prohibition with the given name",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the operations in the prohibition",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "prohibitions"
        ],
        "description": "Add operations to the prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "description": "The request object containing an array of operations to add to the prohibition",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOperationsToProhibitionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The operations were successfully added to the prohibition"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}/operations/{op}": {
      "delete": {
        "tags": [
          "prohibitions"
        ],
        "description": "Delete an operation from the prohibition with the given name",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "op",
            "description": "The operation to delete",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The operation was successfully removed from the prohibition"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}/resources": {
      "get": {
        "tags": [
          "prohibitions"
        ],
        "description": "Get the resources in a prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the resources in a prohibition",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProhibitionResource"
              }
            }
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "prohibitions"
        ],
        "description": "Add a resource to the prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "description": "The request object containing the resource ID and whether to take the compliment of the resource",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddResourceToProhibitionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource was added to the prohibition"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}/resources/{resourceId}": {
      "get": {
        "tags": [
          "prohibitions"
        ],
        "description": "Get the resource identified by {resourceId} in the prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "name": "resourceId",
            "description": "The ID of the resource to get",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A resource in the prohibition",
            "schema": {
              "$ref": "#/definitions/ProhibitionResource"
            }
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "prohibitions"
        ],
        "description": "Delete a resource from a prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "name": "resourceId",
            "description": "The ID of the resource to delete",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The resource was successfully removed from the prohibition"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prohibitions/{prohibitionName}/subject": {
      "get": {
        "tags": [
          "prohibitions"
        ],
        "description": "Get the subject of a prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The subject of the prohibition",
            "schema": {
              "$ref": "#/definitions/ProhibitionSubject"
            }
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "prohibitions"
        ],
        "description": "Set the subject of the prohibition",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "prohibitionName",
            "description": "The name of the prohibition",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "description": "The request object containing the ID of the new subject and the type of the new subject(user, user attribute, or process)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostSubjectToProhibitionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subject was successfully set for the prohibition"
          },
          "400": {
            "description": "ProhibitionDoesNotExistException"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}